+++ packages/manager/apps/dedicated/client/app/app.module.js
@@ -85,6 +85,7 @@ import ovhContacts from '@ovh-ux/ng-ovh-contacts';
 import ovhManagerAccountSidebar from '@ovh-ux/manager-account-sidebar';
 import ovhManagerAtInternetConfiguration from '@ovh-ux/manager-at-internet-configuration';
 import { registerCoreModule } from '@ovh-ux/manager-core';
+import ovhManagerCookiePolicy from '@ovh-ux/manager-cookie-policy';
 import ovhManagerBanner from '@ovh-ux/manager-banner';
 import ovhManagerEnterpriseCloudDatabase from '@ovh-ux/manager-enterprise-cloud-database';
 import ovhManagerDbaasLogs from '@ovh-ux/manager-dbaas-logs';
@@ -141,6 +142,7 @@ export default (containerEl, environment) => {
         account,
         ovhManagerAccountSidebar,
         registerCoreModule(environment),
+        ovhManagerCookiePolicy,
         'Billing',
         chartjs,
         'controllers',